Improve usability for the Facility Edit page.
Disabled scroll-zooming on the map to avoid user frustration when trying
to scroll on the page itself. Resized the map for more optimal page
layout. Repositioned the search box to be located directly above the map
to make it more prominent, also changed search box placeholder text
to be more clear.